requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==21.6
playwright==1.47.0
apscheduler==3.10.4
sqlmodel==0.0.21
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
fast-flights>=2.2
